+++ server/script_packages/py_process.js
@@ -1,9 +1,8 @@
-const { spawn } = require('child_process');
-const JSON5 = require('json5')
+
 const WL_Item = require('../models/WatchListItem');
 const BBItem = require('../models/BBItem');
 const CCItem = require('../models/CCItem');
-const { Product } = require('../models/PriceProduct');
+const {BBScript, BBNumScript, BBSkuItemScript} = require('./Scripts');
 
 const py_process = (_id, link) => {
     //Script Object crawl product price from a link
@@ -34,113 +33,6 @@ const py_clock_cycle = async () => {
             py_process(item._id, item.link);
         }))
     });
-
-}
-// parent BB script 
-class BBScript {
-    constructor(model) {
-        this.model = model;
-        this.script_path = './script_packages/priceTracker.py';
-        this.link = `https://www.bestbuy.com/site/searchpage.jsp?_dyncharset=UTF-8&browsedCategory=pcmcat138500050001&id=pcat17071&iht=n&ks=960&list=y&qp=condition_facet%3DCondition~New%5Eparent_operatingsystem_facet%3DParent%20Operating%20System~Windows&sc=Global&st=categoryid%24pcmcat138500050001&type=page&usc=All%20Categories`
-    }
-    spawnScript(arg) {
-        let arg_strfy = JSON.stringify(arg);
-        console.log(`Start crawling ${this.constructor.name}...`)
-        return spawn('python', [this.script_path, arg_strfy]);
-    }
-
-    spawnPriceScript(_id, link) {
-        const product = new Product(_id, link);
-        const product_arg = JSON.stringify(product);
-        console.log(`Start crawling product price...: ${product_arg}`)
-        return spawn('python', [this.script_path, product_arg]);
-    }
-
-    listenOn(python) {
-        python.stdout.on('data', (data) => {
-            console.log(`Pipe data from script: ${this.constructor.name}...`);
-            this.data = JSON5.parse(data.toString());
-        })
-    }
-    listenClose(python, resolve) {
-        python.on('close', (code) => {
-            console.log(`${this.constructor.name} child process close all stdio with code ${code}`);
-            resolve();
-        })
-    }
-    listenErr(python, reject) {
-        python.on('error', () => {
-            console.log(`${this.constructor.name} child process close with ERROR`);
-            reject();
-        })
-    }
-
-    updateDBPriceById(Model, product) {
-        //update price and name returned from python script, push price_timestamp into price_timestamps
-        Model.findByIdAndUpdate(product.id, {
-            name: product.name,
-            $push: {
-                price_timestamps: {
-                    price: product.currentPrice,
-                }
-            }
-        }, { useFindAndModify: false }, (err, docs) => {
-            if (err) {
-                console.log(`[Error]Update name and price by _id: ${product.id} Failure`)
-            } else {
-                console.log(`Updated _id: ${product.id} Success`)
-            }
-        });
-
-        console.log(`Updated price timestamps, name of product in DB...:\n${JSON5.stringify(product)}`)
-    }
-
-}
-
-class BBNumScript extends BBScript {
-    constructor(model) {
-        super(model);
-        this.script_path = './script_packages/bbLaptopsNum.py';
-    }
-
-}
-
-class BBSkuItemScript extends BBScript {
-    constructor(model) {
-        super(model);
-        this.script_path = './script_packages/bbSkuItem.py';
-    }
-    listenOn(python) {
-        python.stdout.pipe(require('JSONStream').parse()).on('data',(data)=>{
-            // console.log(`Pipe data from script: ${this.constructor.name}...`);
-            console.log(`Pipe data into DB on SKU:${data.sku}\n ${JSON5.stringify(data)}`)
-            this.findSkuAndUpdate(data)
-        })
-    }
-    getLinkInfo(item_num) {
-        return ({
-            link: this.link,
-            link_index: Math.ceil(item_num / 24)
-        })
-    }
-
-    findSkuAndUpdate(item) {
-        let query = { sku: item.sku },
-            update = {
-                name: item.name,
-                link: item.link,
-                price_timestamps: [{
-                    price: item.currentPrice
-                }]
-            },
-            options = { upsert: true, new: true, setDefaultsOnInsert: true, useFindAndModify: false }
-
-        this.model.findOneAndUpdate(query, update, options, (err, doc) => {
-            if (err) return;
-
-            // console.log(`sku-item doc update:${doc}`);
-        })
-    }
 
 }
 
@@ -203,6 +95,35 @@ const py_cc_process = () => {
         console.log("CCNum Script Failure.");
     })
 }
+// get all laptops new condition number promise, resolve when retrieve items number.
+const ccAllLaptopsNewNumPromise = (CCNum) => {
+
+    //spawn script to get items number
+    const python = CCNum.spawnScript(CCNum.link);
+
+    // listen for script, get total items number
+    CCNum.listenOn(python);
+    const getAllLaptopsNumPromise = new Promise((resolve, reject) => {
+        CCNum.listenClose(python, resolve);
+        CCNum.listenErr(python, reject);
+    });
+    return getAllLaptopsNumPromise;
+}
+
+// get all laptops sku-items promise, resolve when retrieve all skus, names, currentPrices.
+const ccAllLaptopsSkuItemsPromise = (CCSkuItems, num_of_pages) => {
+    const link_info = CCSkuItems.getLinkInfo(num_of_pages);
+    const sku_items_python = CCSkuItems.spawnScript(link_info);
+    CCSkuItems.listenOn(sku_items_python);
+    const getCCSkuItemsPromise = new Promise((resolve, reject) => {
+        CCSkuItems.listenClose(sku_items_python, resolve);
+        CCSkuItems.listenErr(sku_items_python, reject);
+    });
+    return getCCSkuItemsPromise;
+}
+
+
+
 module.exports = {
     py_process: py_process,
     py_clock_cycle: py_clock_cycle,
